Move config to config dir, move main to cmd dir, update Dockerfile, update Makefile
