+++ Makefile
@@ -1,6 +1,7 @@
 .PHONY: build version dockerize clean daily-content
 
 version=0.1.0
+app=daily-context
 
 build: daily-content
 
@@ -8,10 +9,10 @@ version:
 	@echo $(version)
 
 dockerize:
-	docker build -t daily-content:$(version) .
+	docker build -t $(app):$(version) .
 
 clean: 
-	rm daily-content
+	rm main-content
 
 daily-content: 
-	CGO_ENABLED=0 go build -o daily-content -mod=vendor *.go
+	CGO_ENABLED=0 go build -v -o main -mod=vendor ./cmd/
